Move PictureView component into ingredients folder
