+++ app/views/alchemy/ingredients/_picture_view.html.erb
@@ -1,4 +1,4 @@
-<%= render Alchemy::PictureView.new(
+<%= render Alchemy::Ingredients::PictureView.new(
   picture_view,
   **local_assigns.slice(:options, :html_options)
 ) -%>
